+++ changelog.md
@@ -1,5 +1,9 @@
 # Gro changelog
 
+## 0.1.2
+
+- add `utils/createObtainable.ts` for decoupled lifecycle management
+
 ## 0.1.1
 
 - add `fs/watchNodeFs.ts` for low level filesystem watching